# src/main/webapp/docs/swagger.yaml
swagger: "2.0"
info:
  description: "An anime api where you can get anime information, photos, videos, news and merchandising. Also you can update your MyAnimeList account."
  version: 1.0.0
  title: "M.A.R.C"
  # put the contact info for your development or API team
  contact:
    email: "grupoproyectomarc@gmail.com"

  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

# all paths will be relative to the basePath specified
# basePath: /v1
# Added by API Auto Mocking Plugin
# host: proyectomarc-161510.appspot.com
# basePath: /api
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /GrupoMARC/M.A.R.C/1.0.0
...
schemes:
  - "https"
paths:
  "/noticias":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias"
      operationId: "getAllNoticias"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todas las noticias"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/idAnime":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias de la id de un determinado anime"
      operationId: "getAllNoticiasById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las noticias de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/titleAnime":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias del nombre de un determinado anime"
      operationId: "getAllNoticiasByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las noticias de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/{id}":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve una noticia con un determinado id"
      operationId: "getNoticiaById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID de la noticia"
          required: true
          type: "string"
      responses:
        200:
          description: "Noticia según un determinado id"
          schema:
            $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/imagenes":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imágenes"
      operationId: "getAllImagenes"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todas las imágenes"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/idAnime":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imágenes de la id de un determinado anime"
      operationId: "getAllImagenesById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las imágenes de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/titleAnime":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imagenes del nombre de un determinado anime"
      operationId: "getAllImagenesByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las imágenes de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/{id}":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve una noticia con un determinado id"
      operationId: "getImagenById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID de la imagen"
          required: true
          type: "string"
      responses:
        200:
          description: "Imagen según un determinado id"
          schema:
            $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
...
  "/perfiles":
    get:
      tags:
      - "Recurso Perfil"
      summary: "Método que devuelve todos los perfiles"
      operationId: "getAllPerfiles"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todos las perfiles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Perfil"
        404:
          description: "Ningún resultado"
    post:
      tags:
      - "Recurso Perfil"
      summary: "Método que añade un perfil"
      operationId: "addPerfil"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "Perfil"
          description: "Campos del perfil"
          required: true
          schema:
            $ref: "#/definitions/Perfil"
      responses:
        201:
          description: "Perfil añadido correctamente"
          schema:
              $ref: "#/definitions/Perfil"
        400:
          description: "Petición incorrecta"
    put:
      tags:
      - "Recurso Perfil"
      summary: "Método que actualiza un perfil"
      operationId: "updatePerfil"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "Perfil"
          description: "Campos del perfil"
          required: true
          schema:
            $ref: "#/definitions/Perfil"
      responses:
        204:
          description: "Perfil actualizado correctamente"
        400:
          description: "Petición incorrecta"
        404:
          description: "Perfil no encontrado"
  "/perfiles/{id}":
    get:
      tags:
      - "Recurso Perfil"
      summary: "Método que devuelve un perfil con un determinado id"
      operationId: "getPerfilById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del perfil"
          required: true
          type: "string"
      responses:
        200:
          description: "Perfil según un determinado id"
          schema:
              $ref: "#/definitions/Perfil"
        404:
          description: "Ningún resultado"
    delete:
      tags:
      - "Recurso Perfil"
      summary: "Método que elimina un perfil"
      operationId: "removePerfil"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del perfil"
          required: true
          type: "string"
      responses:
        204:
          description: "Perfil eliminado correctamente"
        404:
          description: "Perfil no encontrado"
  "/perfiles/usuario":
    get:
      tags:
      - "Recurso Perfil"
      summary: "Método que devuelve el perfil de un determinado usuario"
      operationId: "getPerfilByUser"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "user"
          description: "Usuario del perfil"
          required: true
          type: "string"
      responses:
        200:
          description: "Perfil según un determinado usuario"
          schema:
              $ref: "#/definitions/Perfil"
        404:
          description: "Ningún resultado"
  "/perfiles/anime/{idPerfil}":
    post:
      tags:
      - "Recurso Perfil"
      summary: "Método que añade un anime a un perfil"
      operationId: "addAnimeToMyList"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "MyAnime"
          description: "Campos del anime"
          required: true
          schema:
            $ref: "#/definitions/MyAnime"
        - in: "path"
          name: "idPerfil"
          description: "ID del perfil"
          required: true
          type: "string"
      responses:
        201:
          description: "Anime añadido correctamente"
          schema:
              $ref: "#/definitions/MyAnime"
        400:
          description: "Petición incorrecta"
    put:
      tags:
      - "Recurso Perfil"
      summary: "Método que actualiza el anime de un perfil"
      operationId: "updateAnimeInMyList"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "MyAnime"
          description: "Campos del anime"
          required: true
          schema:
            $ref: "#/definitions/MyAnime"
      responses:
        204:
          description: "Anime actualizado correctamente"
        400:
          description: "Petición incorrecta"
        404:
          description: "Perfil o anime no encontrado"
  "/perfiles/anime/{idPerfil}/{idAnime}":
    delete:
      tags:
      - "Recurso Perfil"
      summary: "Método que elimina un anime de un perfil"
      operationId: "deleteAnimeInMyList"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "idPerfil"
          description: "ID del perfil"
          required: true
          type: "string"
        - in: "path"
          name: "idAnime"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        204:
          description: "Anime eliminado correctamente"
        404:
          description: "Perfil o anime no encontrado" 
...
  "/animes/{id}":
    get:
      tags:
      - "Recurso Anime"
      summary: "Método que devuelve un anime con un determinado id"
      operationId: "getAnimeById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Anime según un determinado id"
          schema:
              $ref: "#/definitions/Anime"
        404:
          description: "Ningún resultado"
    delete:
      tags:
      - "Recurso Anime"
      summary: "Método que elimina un anime"
      operationId: "removeAnime"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        204:
          description: "Anime eliminado correctamente"
        404:
          description: "Anime no encontrado"
...
definitions:
  Anime:
    required:
      - id
      - name
      - episodes
      - startDate
      - endDate
      - synopsis
    properties:
      id:
        type: string
        example: a1
      name:
        type: string
        example: Higurashi No Naku Kori ni
      episodes:
        type: string
        example: 26
      startDate:
        type: string
        example: 05/04/2006
      endDate:
        type: string
        example: 27/09/2006
      synopsis:
        type: string
        example: Keiichi Maebara has just moved to the quiet little village of Hinamizawa in the summer of 1983, and quickly becomes inseparable friends with schoolmates Rena Ryuuguu, Mion Sonozaki, Satoko Houjou, and Rika Furude. However, darkness lurks underneath the seemingly idyllic life they lead.
      imagenes:
        $ref: '#/definitions/Imagen'
      noticias:
        $ref: '#/definitions/Noticia'
      productos:
        $ref: '#/definitions/Producto'
      videos:
        $ref: '#/definitions/Video'
  Imagen:
    required:
      - id
      - size
      - url
      - title
    properties:
      id:
        type: string
        example: i1
      size:
        type: string
        example: 1080x920
      url:
        type: string
        example: https://68.media.tumblr.com/c705d57e15bb02a9bbcfe2f49e2267be/tumblr_oq2vo67Zm61t6nfoqo1_1280.jpg
      title:
        type: string
        example: Cosplay higurashi
  Noticia:
    required:
      - id
      - title
      - url
      - content
    properties:
      id:
        type: string
        example: n1
      title:
        type: string
        example: New promo
      url:
        type: string
        example: http://www.animenewsnetwork.com/news/2011-07-21/higurashi-no-naku-koro-ni-kira-magical-girl-promo-posted
      content:
        type: string
        example: Animate began streaming a "Magical Girl" promotional video for the Higurashi no Naku Koro ni Kira original video anime (OVA) series on Thursday.
  Producto:
    required:
      - id
      - name
      - url
      - prize
      - description
    properties:
      id:
        type: string
        example: p1
      name:
        type: string
        example:  Cosplay Rika
      url:
        type: string
        example: https://es.aliexpress.com/item/Higurashi-no-Naku-Koro-ni-Higurashi-When-They-Cry-Rika-Furude-red-Witch-kimono-Cosplay-Costume/32350149865.html?spm=2114.43010208.4.1.P4V9SM
      prize:
        type: string
        example: 57,68 € 
      description:
        type: string
        example: Higurashi no Naku Koro ni Higurashi When They Cry Rika Furude red Witch kimono Cosplay Costume anysize
  Video:
    required:
      - id
      - title
      - url
    properties:
      id:
        type: string
        example: v1
      title:
        type: string
        example:  Higurashi no Naku Koro ni by Eiko Shimamiya
      url:
        type: string
        example: https://www.youtube.com/watch?v=XgWkIrFQQE0
  Perfil:
    required:
      - id
      - nick
    properties:
      id:
        type: string
        example: pe1
      nick:
        type: string
        example: Raziel
      myAnimes:
          $ref: '#/definitions/MyAnime'
  MyAnime:
    required:
      - id
      - name
      - episodes
      - startDate
      - endDate
      - synopsis
      - myScore
      - myStatus
    properties:
      id:
        type: string
        example: a1
      name:
        type: string
        example: Higurashi No Naku Kori ni
      episodes:
        type: string
        example: 26
      startDate:
        type: string
        example: 05/04/2006
      endDate:
        type: string
        example: 27/09/2006
      synopsis:
        type: string
        example: Keiichi Maebara has just moved to the quiet little village of Hinamizawa in the summer of 1983, and quickly becomes inseparable friends with schoolmates Rena Ryuuguu, Mion Sonozaki, Satoko Houjou, and Rika Furude. However, darkness lurks underneath the seemingly idyllic life they lead.
      imagenes:
        $ref: '#/definitions/Imagen'
      noticias:
        $ref: '#/definitions/Noticia'
      productos:
        $ref: '#/definitions/Producto'
      videos:
        $ref: '#/definitions/Video'
      myScore:
        type: string
        example: 7
      myStatus:
        type: string
        example: On hold